Refactor setFireDuration to accept int duration
Updated the setFireDuration method to use int instead of float for duration.

+++ src/main/java/com/onyxi7/betterarchery/compat/crafttweaker/CTArrows.java
@@ -7,14 +7,6 @@ import crafttweaker.annotations.ZenRegister;
 import stanhebben.zenscript.annotations.ZenClass;
 import stanhebben.zenscript.annotations.ZenMethod;
 
-/**
- * CraftTweaker integration for Better Archery arrows.
- * Allows modpack makers to customize arrow stats.
- * 
- * Example usage in .zs script:
- *   mods.betterarchery.Arrows.setDamageMultiplier("betterarchery:fire_arrow", 1.5);
- *   mods.betterarchery.Arrows.setExplosionRadius("betterarchery:impact_arrow", 3.0);
- */
 @ZenClass("mods.betterarchery.Arrows")
 @ZenRegister
 @ModOnly("betterarchery")
@@ -35,7 +27,7 @@ public class CTArrows {
     @ZenMethod
     public static void setFireDuration(String arrowRegistryName, int duration) {
         CraftTweakerAPI.logInfo("[BetterArchery] Setting fire duration for " + arrowRegistryName + " to " + duration);
-        ModpackOverrides.arrowFireDurations.put(arrowRegistryName, (float) duration);
+        ModpackOverrides.arrowFireDurations.put(arrowRegistryName, duration);
     }
     
     @ZenMethod
